Some changes, mostly to tables…
Table-heads without table-header sort-option do not have a hover-effect
any more, simplified table layout for that change; rebuild some forms to
be built upon definition lists; extended the CSS for large buttons;
changed some CSS for the page-background and its effects (unfinished).
Note that background-colors on page-back, box-titles and table-heads are
dark, because there should be a background-image over that. Images
aren't checked in at this time. Maybe the design will change to the
"Basic"-style at some point.