Fixed 2 errors that prevented the installation of standalone application
`PackageInstallationNodeBuilder::buildPackageNode()` didn't handle
non-existing `packageDescription` and `authorURL` tags (which may be
omitted according to the `package.xsd` file) so that they were set to
null which caused problems later when
`PackageInstallationDispatcher::installPackage()` is called and tries to
save that node data since the database table columns
`packageDescription` and `authorURL` may not be null.  Also the
`SQLPackageInstallationPlugin::install()` method called
`Package::getAbbreviation()` the wrong way.