Fixed threw package update errors
Those three errors are: 1. After a package is updated, the package
version isn't updated. 2. Several pieces of code caused an error in
`PackageInstallationNodeBuilder::buildPluginNodes()` because
`PackageArchive::getUpdateInstructions()` returned an array with an
illegal structure that was created while building the update
instructions. 3. `use wcf\util\StringUtil;` was missing in `Package`
since the `StringUtil` is used in `Package::checkFromversion()` (just
used during updates, not during installations).