Fixed various errors concerning cronjobs
1. The default value of the state column of the cronjob database table
has to be 0 rather than 1 since 1 stands for pending and 0 for ready.
Currently no cronjob will ever be executed. 2. The method
`ICronjob::execute()` expected an array with the cronjob data. First of
all, the `CronjobScheduler` called the method without any argument which
caused an error. Since we now have a DatabaseObject for cronjob that is
available when the method is called, I changed the parameter from being
an array to a Cronjob DatabaseObject. 3. I changed the implementation of
`CronjobScheduler` to a Singelton/subclass of `SingeltonFactory` and
changed the name of the execution method from `execute()` to
`executeCronjobs()` since the the cronjobs are executed rather than the
scheduler. 4. Fixed a wrong namespace in `CronjobScheduler` and some
misnamed variables. 5. I changed "classPath" to "className" in the
cronjob log page. 6. "Disabled" `GetUpdateInfoCronjob` and
`RefreshSearchRobotsCronjob` for the moment until the needed fixes are
done so that these cronjobs don't cause any errors during execution. 6.
Added the missing length of the success database table column of
cronjob_log (1, since just 1 and 0 are possible success values). 7.
Fixed database table name in `CleanUpCronjobLogCronjob`. 8. Added
missing update of the last execution time of the cronjob. 9. Fixed wrong
file name in `CacheScheduler::clearCache()`. 10. Removed unnecessary
LEFT JOIN `CacheScheduler::loadCronjobs()` to get package.packageDir
since the the class is autoloader via it's namespace. 11. Updated the
documentation a lit.