Added the method AbstractPackageInstallationPlugin::getDatabaseTableName()
This method can easily be overwritten by package installation plugins
that are bought with a standalone application (and then, of course, only
work within it), like a pip to install new items for a menu of the
standalone application. Without that method, each other method that uses
`…wcf".WCF_N."…` in queries would have to be overwritten.