More SVG-icons and their implementation in the templates, corrected a mistake in 2 comments, some minor CSS changes.
1) New icon-naming system: At this time icons without trailing digit are
meant to be blue (imaging a "0" as number), those with a "1" are
supposed to be dark grey and those with a "2" are supposed to be white.
An icon with an additional "D" in its name has less opacity (regardless
of its color) and be nearly transparent. Note that this icon-naming
system may not be final and change over time. Note that due to the
possibilities of SVG-images and CSS3 techniques, there is no need for
icons in different sizes any more, which reduces the number of icons
dramatically.   2) If we find a way to put multiple icons in one file,
and/or multiple states of them, there will be even less icons. This is
just the state of establishing SVG icons generally, it may become more
smart in future.