Fixed calculation of afterNextExec
The timebase for calculation of `afterNextExec` is equal to `nextExec` in order to find any date after this time for execution. Unfortunately this timebase is already a valid date and thus is used as result (now `nextExec` and `afterNextExec` are equal). To fix this, we simply add 1 to the `nextExec` timestamp, making it an invalid date for next execution (we do not support seconds anyway!).