Again table-views & some minor stuff
Unified table views (orientations & column CSS classes); Extended the
set of standard-classes for different table-row designs, they are
always the first declaration, please use those classes only (but note,
that in some cases there are generic table-columns, which cannot be
styled individually that way, we will have to add the generic classes
to the matching CSS declaration from time to time); Changed the way
elements on pages hold their distance to others. Note that we now use
only margins to the top, but no margins to the bottom any more; Edited
some cronjob texts; Edited the worker-template to reflect the changes
in the other overlay-templates.