Small corrections
We need a way to show a hand-cursor on every element which commits an
action, such as delete or info, not depending if there's a link on that
element. So that should work on plain images with Javasript-actions
too. This should not de related to the balloon-tooltip; small change in
package list.