Implemented a debug switch
It is now possible to switch between debug and production, whereas the last setting will suppress any error message including file/line references and stacktraces.

Furthermore errors will ALWAYS be written to a logfile, regardless if running in production or debug mode.

By default debug mode is enabled for now, it MUST be set to 'production' once the software will be tossed into the wild.