Fixes a possible bug in AbstractDatabaseObjectAction::update()
This bug currently occurs if you try to edit the profile inline in the profile itself (Error: "Undefined index: data"). The reason for this bug is that the after saving the input data, the "real" user data isn't updated, only their options which is handled by the UserAction separately (UserAction, line 140), so 'data' doesn't always need to exist.