Adds default titles for toggle buttons in javascript handler
If there is no specific title for the enable/disabled button is given, the neutral/general title is used.

"disableMessage"/"enableMessage" is replaced by "disableTitle"/"enableTitle" since that is - in my opinion - more accurate because the title is simply changed and you don't see any pop-up message appearing after changing the status.