Introducing the CSS-class "button"
The class "button" will now be uset to give button-type elements their
color and style, while the size and overall look will be determined by
an additional class like "largeButtons". This makes transforming
elements to a button more easy. This works mostly now, but there are
some elements that need to be fixed still! The style changes will be
committed later …