Prefixes for classes; New icon; Optimizations
Experimental: All classes will be prefixed from now on. This is the
start of this experiment. Prevents conflicts with
3rd-party-stylesheets. At the moment all classes are being doubled for
compatibility reasons, i'll remove those duplicates when all classes
are changed. Some problems: What will be the prefix for classes, which
are not WCF- or WBB-depended? Styles coming from small packages should
be prefixed somehow.