Introducing the "container"-class
This is a new generic container for surrounding blocks of text with an
optional leading icon, which helps to have less style-declarations. The
good thing: Icon-size is free. The box collapses if there's no icon
given or  if it doesn't load. The class "containerIcon" can be set onto
an icon or an element around the icon.

Note: This box does unfortunately not support RTL in WebKit-browsers!
This seems to be a bug in WebKit, since all other browsers support
that. The box-style may change in the future.

Implemented the new classes on main-headings.