Fixes a bug for application with no lang vars
This bug should also occur if the active application supports less languages than installed and the installed package supports languages that are installed but not supported by the active application.

In those cases, the localized package infos weren't saved since the language array way empty (or not all localizations were saved). Now all languages are fetched and now this also happens at the very end of the installation which makes it possible to save the localized data in a new language that was delivered by the installed package (though this seems to be not supported yet).