Adds the possibility to auto-trigger action effects in WCF.Clipboard
If a clipboard item has the following parameters:

* actionName
* className
* objectIDs (ids of the marked objects which would be affected by the action)

the following things happen:

* An AJAX action with actionName and className is executed (if additionally a confirm message is given, the user has to handle that first before the action is executed).
* It's checked if an action object for the relevant object type and actionName exists in WCF.Clipboard exists and for each object id in objectIDs triggerEffect(objectID) on the action object is called. This way deleted objects are also removed from the table or toggled objects get updated toggle "buttons".

To achieve the second part, WCF.Action.Delete and WCF.Action.Toggle have a new method: triggerEffect() which is basically the content of _success but can now be called from WCF.Clipboard with the relevant object ids.
Furthermore UserClipboardAction returns the relevant user ids if the delete action is active.

---

Additionally, I update the javascript handling of the user list in the acp:

* The deletion button works and uses WCF.Action.Delete.
* The new WCF.Table.EmptyTablerHandler is in use.
* The above mentioned WCF.Clipboard changed are implemented which is mainly adding a few javascript lines and removing several obsolete lines from WCF.ACP.