Makes it easier to use own category nodes
With the changes of this commit you can easily extend the given node classes (without having to copy&paste the `__construct` methods and just change the class name). This makes it possible for your own implementation to add additional conditions for a category to be included in the list (like acl options "canViewCategory" or "canViewDisabledCategory").