Fixes erroneous clipboard behaviour related to type data
Previously, the type data wasn't considered when the editor label was fetched which could result in a higher number of marked items (since all marked items of the relevant object type were used). This commit fixes this in such way that avoids multiple checks. Before anything else, the objects are filtered by the clipboard action handler *once* using the type data, and then given to execute() and getEditorLabel() where no filtering will be necessary.

Changes in other classes because of this bug fix are minimal: If you don't support type data, simple extend AbstractClipboardAction instead of implementing IClipboardAction and remove the third parameter of execute() [since it's optional, keeping will not break anything, but keeping it there makes no sense].