Tries to fix ACP menu again
Those top option category links are trying really hard to make one's life miserable.

Previously, the session ID of the active user was cached in the top option category links, which will break them in other sessions/for other users. Now, since these links are different and included in the core, `ACPMenuItem` is allowed to know that there are such "special" links and this it can handle them by building the link with the id of the option category which is cached with them as a "special" value.