Date Picker now supports min/max dates and RTL mode
I haven't found a var in the WCF indicating the first day of the week. Maybe you want to add a new language var for this, but it's not that easy: Most countries and ISO 8601 say it's monday (new default), the US says it's sunday (old default). But: UK says monday, too - so it's not really language-specific. Maybe according to the time zone? Do what ever you prefer...