Allow breadcrumbs without an link
This should be avoided wherever possible, but sometimes there are good reasons for a breadcrumb without an link. For example, when using sidebars, you probably have a navigation like
* Category
* * Item 1
* * Item 2

Usually "Category" doesn't represent a page, it's just a logical container for its containing items. Adding the category to the breadcrumbs navigation results in a much better user experience. When the user navigates deeper, the resulting breadcrumbs navigation would be
Homepage > Category > Item 1
Homepage > Category > Item 2

instead of just
Homepage > Item 1
Homepage > Item 2

This makes the user easier to understand that "Item 1" and "Item 2" are of the same kind.