Adding the ability to bind objects to callbacks
Unfortunately the "real" support of this was added in PHP 5.4, so we can't use it. This is a way to work around that pitfall. Please note that function() use ($this) {} doesn't work in PHP 5.3. This suggestion is the same as doing
    $callback = function($self) { return $self->property; };
    $callback($this);
but I think this is a nicer way of doing that - and that's actually the only reason for this Callback class, isn't it?