WCF.Date.Time now properly handles time offsets
The calculation of relative times previously failed if both the client and the server use the timezone, but their actual time is slightly different. In that case both times could be different by a few seconds, causing times to be in the future, even though it was just a mismatch.

Fixes #1179