Adds change event trigger for ColorPicker if color is updated
If you want to use the color picker to select a color for an element that is currently visible, it might be nice to update the color of the element "live", i.e. each time the user updates the selected color, the element gets the new color. Currently, there is no way to detect when the user has changed the selected color. Now, they can simply listen to the `change` event of the input field.