Work-around for broken APC versions
APC up to 3.1.3p1 breaks PHP's late static binding, the only way to support it is constantly clearing the opcode cache. This isn't the best solution, but it works out and does not prevent the user from using our software.

Sadly 3.1.3p1 (the last broken version) is the stable release available on Debian old-stable "Squeeze" :(