Added dereferrer
A dereferrer (like the one, which is used for acp) is a good idea.
People will be able to track and limit outgoing links, hide referrer or
just display some kind of a disclaimer for outgoing links, etc.
Currently, there's no control over external links. The only way to
manipulate them is JS, which is not the best idea in this case.