Removed transition between content and sidebar for mobile devices
It is not possible to use a transition in Internet Explorer unless we're using floats which in fact suck. width: 200% for display:table results in the strangest thing I've ever seen and furthermore that browser doesn't give a damn about white-space: nowrap with two inline-block children.

Thank you Internet Explorer.

Not.