Fixes languageCustomItemValue for "reseted" language items
If you use a custom version of language item, but don't want to use it anymore and don't want the custom version to be saved, you uncheck the "Use custom value" checkbox and leave the "Custom Value" textarea empty. This is the same state as the state of a never edited language item which is why the `languageCustomItemValue` value of the previously customized language item has to be `NULL` as it is for the never edited language item.

This fixes the number of custom phrases in the language table of LanguageListPage which currently counts all language items which ever had a custom value.