Fixes counter update when deleting comment with responses
When deleting comments with responses, the object type processor is first called with the number of responses plus one (for the deleted comment itself). This call correctly corrects the number of comments for the relevant object.

Afterwards, when deleting the responses, the objects' comment counters were updated again for each deleted response which causes a incorrect count because the counters were now updated twice for each response. Thus, we can simply ignore the counter update altogether in CommentResponseAction when deleting responses because the comment, they belong to, is deleted.

Furthermore, there is no need for CommentResponseAction to update the response counter of those comments in this case because the comments will be deleted anyway.