Added delay for search area
Previously a request would be send for every key stroke, now a request will only be send, if there was no key stroke for 500 miliseconds. This means that if you type pretty fast, it will only trigger a search once you stop.

This should reduce "hammering" the server with request which are more or less void since they will be ignored.