Improved Redactor integration
The biggest change was the transition to use <p> instead of <br> (disabled linebreaks mode), because attempts to style a lonely textnode (with no actual parent element except the editor container) causes a lot of issues.

Furthermore I have started to modify the HTML of Redactor's modal overlays, they (partially) mimic the appearance of WCF's overlays.