Combining redactor plugins into a single, compressed file
Right now there are 8 plugins so far, this would normally cause 8 HTTP requests just to fetch them or check if the cached file is still up-to-date. Since it does not matter in which order these plugins are fetched (the editor executes them at a later point in the correct order), we can safely combine them into a single, compressed file. This saves us 7 requests. Glorious!