Changed JavaScript execution order
This way we ensure that specific JavaScript modifications to the DOM (e.g. relative time or TabMenus) are executed as early as possible. Tests have shown that there is no more flicker caused by the time conversion or building the TabMenu, in general the related JavaScript is executed earlier (reduction by ~20%)

The impact on mobile devices is low and the small delay due to earlier loading of certain JavaScript files trades for a better user-experience, especially on small screens the "jumping" of content due to initial DOM manipulations was pretty lame.

Affected DOM manipulations:
 - Relative Timestamps
 - Dropdowns (e.g. mobile navigation)
 - Tab Menus
 - Flexible Tab Menus (hiding tabs in a dropdown to prevent scrolling/overflow, depends on Dropdowns)