Preventing deletion of delivered BBCodes, Page Menu Items and User Options
All three types are either complex objects or have data or implicit properties associated with them which are not alterable through the ACP. For example user-created menu items will be assigned a generic name which cannot be used as a replacement for pre-defined items with a specific name (e.g. highliting the active menu item).

In order to prevent users from accidentially breaking stuff, we disallow deletion. In return all three types support disabling of items, therefore they can be safely turned off which in return results in the same behavior compared to deleting them.