Deprecating the VersionableDatabaseObject-system
The entire design is flawed as it is anything but flexible and forces copying of full table contents rather than being able to specify the columns which should actually be tracked. In addition it does not provide centralized access and management including comparisons.

The entire API is now deprecated and will be removed with WCF 2.2.