Fix mention issue
Whenever the users presses any key while typing, we have to abort previous requests to make sure that they do not mess up anything.
We also need to consider requests which are not send at the moment but will be send in a moment because of the delay introduced by WCF.PeriodicalExecuter by stoping the periodical executer.