Rename \wcf\system\event\listener\IEventListener to IParameterizedEventListener
This is for BC reasons, as PHP does not allow using a class of another
namespace when a class with the same name already exists within the same
namespace. IParameterizedEventListener will be aliased to IEventListener
when \wcf\system\event\IEventListener is removed.

see bd8f17f01975dacba19199ac4b7f4bba567f3e98