Fix connection failure handling in MemcachedCacheSource
The previous test did not work at all, as `get` does not error out
if the connection fails, but rather returns that the key is not
available.

`set` will return false if the item could not be saved. In addition
\Memcached::OPT_REMOVE_DAFILED_SERVERS is set, which will cause the
broken memcached servers to be ignored for the rest of the execution.
The keys intended for the broken server will transparently be mapped
to other memcached servers.