Allowing specification of referenced object id
Notifications for liked objects suffer from the objectID column referencing the likeID, but this creates a situation where identifying notifications for a certain object (e.g. a post) require an additional roundtrip.

The baseObjectID is used nowhere within the notification system, you can store anything you want and still be able to programmatically retrieve notifications based on it.