Fallback to base64-encoded favicon for multiple domain setups
Favico.js uses a <canvas> to draw the badge, but this will fail if the accessed domain is foo.example.com and the favicon was served from bar.example.com. This is caused by the fact that the <canvas> is considered tainted once an external image has been loaded and thus refuses to export the resulting image.

Roughly 99.99% of all setups use only a single domain, thus there is absolutely no change in this case. This change adresses potentially issues when running a multiple-domain setup.