Using JavaScript to detect iOS devices
The "overflow: hidden" is required to fix the famous zoom bug on orientation change in iOS Safari, sadly the work-around using the event listener for device acceleration is not always reliable.

The previous code caused Chrome on Android to block scrolling if a visible element with "position: fixed" is present (for whatever reason, enlight me Google!), which is for example the case if the quote overlay is present.