Prevent a weird issue in Chrome for Android
Chrome for Android somtimes report an empty value for a text input if the previously typed-in character is an underscore. For some reason the event is not properly bound to the original "this"-context, yielding unexpected results. To clarify this, I have no idea why it happens, thus I simply work-around it and ignore it for the time being.