Use atomic group to match quotes in Highlighter.class.php
This modification is safe, as backtracking the contents of the
parentheses will not help finding a match anyway, as the part
between the parentheses cannot match the closing quote character.

Previously PCRE would have build up a huge stack of backtracking
states for strings containing a single quote character and a large
number of other characters, leading to stack overflow and SIGSEGV
as the final result.