Improved pasting of content during WYSIWYG mode
This of course fails if a page has a completely messed up markup e.g. Spiegel Online, but I cannot fix the view for them without messing up the other. The current approach deals best with the average content encountered, if it encounters messy markup it follows the simple rule: Garbage in, Garbage out.