Rebinding event handlers after certain actions took place
Undo/Redo are two examples which cause issues because they internally work by saving the DOM and replacing it with the stored content. Given the fact that this works on a string, all event handlers are lost and must be reattached once the action has completed.

Luckily we can hook into this.observe.load() which is always invoked after such an action took place.