Providing a handler to adjust cursor position on quote space click
If you click on the margin created by quotes, browsers redirect the caret inside the quote. This can make it impossible to get the caret in front of/after it if there are no siblings. Even though you can use the arrow keys to navigate there, this is rather unknown to users and does not solve the issue on mobile devices.

This change attempts to detect clicks into the margin by comparing the offsets with the click position, ultimately moving the caret where the user excepts it to be.