Improved quote selection for messages containing quotes
Quotes are no longer considered to be valid and are removed from the selection and virtual message body. In addition quotes cannot be created if the mousedown or mouseup occured within a quote, simply checking with indexOf() does not work if the same exact phrase appears outside the quote but within the message body.

Quoting quotes isn't great because the system looks at it on a text level where it cannot understand that the quoted text belongs to a completely different message, causing confusion because the to be created quote will reference the message author.