Permission to manage paid subscriptions no longer depends on the module
The module dependency caused the permission to be discarded if the module is disabled and the target group has been saved meanwhile. This is a side-effect caused by the way the option handler deals with non-present options (and even further with checkboxes in general), but is causing too much confusion.

It is better to properly hide these permissions, but unless we redesign parts of the option system to better handle these cases, hiding this single option is not worth the actual downsides caused by it.