Initial commit for WCF 2.2 w/ require.js
This is a highly experimental version and it is neither complete nor fully working. You cannot even install it at this time, if you really must play around with it, use version 2.1 and copy the necessary changes/additions.

At this point it is not meant to be actively development in terms of new features or API changes, I'm solely focusing on playing around with require.js and see how these things turn out.