Strip newlines when compiling if conditions in templates
Previously, the compiler could handle most situations with newlines but
had issues in certain situations with closing parentheses being followed
by a newline. This case is relevant when code of multiple template listeners
is inserted at the end of an if condition as the template engine concatenates
template listener code by newlines.