Boolean user group options now support 'Never'
The boolean option type allowed only for a simple true/false, which is sometimes quite difficult to work with. For example revoking group permissions is usually accomplished by fiddling with all groups.

The 'Never' option solves this by using the unique value '-1' to revoke a permission regardless how often it was granted.