Work-around for description in non-i18n environments
If there is only a single language installed, the actual content is the
actual content and not a phrase. The `__optional` attribute will cause
`{lang}` to not yield any content at all.

This is pretty much an edge case because we want to suppress
non-existing descriptions to be displayed as a generic phrase.