Removed php-di
​Originally intended to phase out SingletonFactory and get rid of its
downsides, there are a few issues that effectively prevent us from using
DI right now:

* There are a rather large amount of current singletons that impose a
circular dependency on runtime, resolving these would require more
changes to the API and the behavior than we're comfortable with
* Being able to replace classes with a derived one is a great feature,
but unfortunately the event system isn't exactly aware of this and
unless one explicitly opts-in for inheritance, event listeners would not
be executed. This could lead to rather unpredictable behaviors and
debugging issues could be rather challenging
* Only some classes qualify for DI, for instance SingletonFactory has a
locked constructor that can be overriden without breaking BC. This leads
to a mixture of "real" DI classes (using constructor injection) and
others that explicitly invoke the DI container to fetch dependencies on
runtime, negating one key aspect of DI (knowledge about dependencies w/o
examining the class source)

These are the major issues with the current implementation and we rather
chose to ditch DI for now until we have proper solutions for all issues.
Most likely this will mean that we won't see DI in the 2.x tree of WCF
at all.