Using a matching line-height for icons
Previously the line-height was set to 1em which leads to a distinct
misalignment due to the icon being smaller than the fixed dimensions
element.

Increasing the line-height to match the vertical size causes the icon to
be aligned with the base-line where it previously was slightly shifted
to the top of the element.