GD extension is now a requirement
All though stated as optional for a long time, it should have always
been a requirement. The optionally supported ImageMagick extension does
not offer replacement functions for core functions such as
`getimagesize()` which is frequently used to determine images dimensions
and type. This function (among others) aren't part of the PHP core, but
instead belong to the GD extension itself.

This change doesn't eliminate the ability to use ImageMagick to off-load
heavy work, but instead ensures that GD is available due to being a
crucial component.