Ability to assign custom search dropdown label
This label is verified through JavaScript, but can be used to set a
custom label that is actually valid but represents a parameter-based
subtype of an existing type.

For instance Burning Board has the type 'forum' but also special
subtypes to search a thread/board, but this causes the label to read
'forum' until the JS kicks in and changes it if necessary. Statically
setting the correct label ahead of time prevents flickering.