Change type of wcf1_background_job.job to MEDIUMBLOB
When developing naturally some jobs tend to fail due to a bug,
usually one wants to reset the time to allow an immediate execution
once the bug is fixed. With a MEDIUMTEXT column however it is too
easy to accidentally destroy the whole job, as the serialized data
structure contains null bytes that will get mangled when not paying
attention.