easy way to disable WCF's branding via AbstractPage
implemented the boolean `\wcf\page\AbstractPage::$showWoltLabBranding`.

This change is needed to provide an easy way (for 3rdparty-developers)
to hide WoltLab's copyright notice within the footer on specific pages
by setting it to `false`.
This gives every developer the possibility to show a specific copyright
without showing the general one. So far so good we could also do this by
setting `WOLTLAB_BRANDING` to `false`. But setting this option would
disable WoltLab's copyright at all which is not the wanted effect in
most cases and would cause much trouble in case WoltLab's products are
installed and no copyright-free-license has been purchased.
Maybe it would even more senseful to remove `WOLTLAB_BRANDING` at all
and use `AbstractPage::$showWoltLabBranding` only.

In a nutshell: Without such a solution there is no easy way to disably
only WoltLab's copyright on only a few pages.