Speed up parseSmilies in MessageParser
If the WCF holds many smilies (thousands, e.g. due to the emoji plugin by SoftCreatR), the loop in the parseSmilies function needs a significant amount of time. Checking beforehand with strpos and only calling preg_replace, if the smiley was found at least one time, leads to an improved performance.