use meta data before searching page by controller
`self::getActiveRequest()->getPageID() ?: null;` will check the meta data for a page id before it searches the first match of the controller's classname.
This is also used by the BoxHandler, so it seems to be inconsistent in it's handling. In some special cases (in my case it's a Fireball-Page) the BoyHandler will return a valid page id and assigns the boxes correctly while this method returns null or the first match which is wrong in the most cases.
This change would improve the page-mapping and should not have any unwanted side-effects.

You also could shorten the whole method to
```
if (self::getActiveRequest() === null) {
	return null;
}

return self::getActiveRequest()->getPageID() ?: null;
```
because it covers both look-ups within the method (before this commit)